+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 10.9.0
+- [feature] Added `limitedUseToken(completion:)` for obtaining limited-use tokens for
+  protecting non-Firebase backends. (#11086)
+
 # 9.5.0
 - [added] DeviceCheck and App Attest providers are supported by watchOS 9.0+. (#10094, #10098)
 - [added] App Attest provider availability updated to support tvOS 15.0+. (#10093)